\name{survos-package}
\alias{survos-package}
\alias{survos}
\docType{package}
\title{
\packageTitle{survos}
}
\description{
\packageDescription{survos}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{survos}
\packageIndices{survos}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{survos}

Maintainer: \packageMaintainer{survos}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
